+++ NEWS.md
@@ -1,5 +1,9 @@
 # era 0.4.1
 
+This patch fixes a malfunctioning test that was causing CRAN checks to fail. There are no significant changes.
+
+* Fixed malfunctioning check
+
 # era 0.4.0
 
 * New eras: